Feat entity highlighter (#78)
* Created a "TestingPlugin" that allows me to dump data from the game into my console for other plugin development.

Created an "EntityHighlight" plugin that allows the user to put a label over entities as they come across them in the world.  This will be configurable based on entity name.  (Right click name)

Add in an option for a Button to be added into a plugin settings window.

* Fixed culling of divs when the entity is no longer in the EntityManager. (Turns out things aren't rendered but are still in the EntityManager for a very long time)

Add in configurable Entity Highlighting, the list is comma seperated. (Tree,Bank Chest, Water Obelisk)

Removed some unneeded code (My Testing plugin and the references to said plugin.)

* Updated "SettingsType.Range" to have a min and max that are optional.

* Added in more configuration options
 * Offset adjustment
 * Color changing
 * Alpha changing

* Removed Min/Max from SettingsType
Updated EntityHighlight to use validation to clamp the number between 0 and 10.

---------

Co-authored-by: KKonaOG <38236274+KKonaOG@users.noreply.github.com>